+++ NEWS.md
@@ -1,5 +1,8 @@
 # Rsequoia2 (development version)
 
+## v0.0.3
+- Add custom formula capabilities in total row of synthese tables
+
 ## v0.0.2
 
 ### Added